Mostrar mensajes de confirmación y espera - SweetAlert2

+++ src/modules/daybook/views/EntryView.vue
@@ -50,6 +50,7 @@
 <script>
 import { defineAsyncComponent } from 'vue'
 import { mapGetters, mapActions } from 'vuex' //computed
+import Swal from 'sweetalert2'
 
 import getDayMonthYear from '../helpers/getDayMonthYear'
 
@@ -105,6 +106,12 @@ export default {
         },
         async saveEntry(){
 
+            new Swal({
+                title: 'Espere por favor',
+                allowOutsideClick: false,
+            })
+            Swal.showLoading()
+
             if( this.entry.id ){
                 //Actulizar la entrada
                 await this.updateEntry(this.entry)
@@ -115,11 +122,32 @@ export default {
                 this.$router.push({ name:'entry', params: { id }  })
             }
 
+            Swal.fire('Guardando', 'Entrada registrada con éxito', 'success')
+
 
         },
         async onDeleteEntry(){
-            await this.deleteEntry(this.entry.id)
-            this.$router.push({ name:'no-entry' })
+
+            const {isConfirmed} = await Swal.fire({
+                title: '¿Está seguro?',
+                text:'Una vez borrado, no se puede recuperar',
+                showDenyButton: true,
+                confirmButtonText:'Si, estoy seguro'
+            })
+            if(isConfirmed){
+                new Swal({
+                    title:'Espere por favor',
+                    allowOutsideClick: false
+                })
+                Swal.showLoading()
+
+                await this.deleteEntry(this.entry.id)
+                this.$router.push({ name:'no-entry' })
+
+                Swal.fire('Eliminado', '', 'success')
+            }
+
+            
         }
     },
     created(){
